extract tree item getter
A function is extracted that generates the tree label components.
